PitchBendControl: Extract PitchBendControl from RateControl.
