+++ src/engine/controls/ratecontrol.h
@@ -6,6 +6,7 @@
 #include "engine/controls/enginecontrol.h"
 #include "engine/sync/syncable.h"
 
+class PitchBendControl;
 class BpmControl;
 class Rotary;
 class ControlTTRotary;
@@ -24,22 +25,8 @@ public:
   RateControl(const QString& group, UserSettingsPointer pConfig);
   ~RateControl() override;
 
-  // Enumerations which hold the state of the pitchbend buttons.
-  // These enumerations can be used like a bitmask.
-  enum class RampDirection {
-      None,      // No buttons are held down
-      Down,      // Down button is being held
-      Up,        // Up button is being held
-      DownSmall, // DownSmall button is being held
-      UpSmall,   // UpSmall button is being held
-  };
-
-  enum class RampMode {
-      Stepping = 0, // pitch takes a temporary step up/down a certain amount
-      Linear = 1    // pitch moves up/down in a progressively linear fashion
-  };
-
-  void setBpmControl(BpmControl* bpmcontrol);
+  void setPitchBendControl(PitchBendControl* pitchBendControl);
+  void setBpmControl(BpmControl* bpmControl);
 
   // Returns the current engine rate.  "reportScratching" is used to tell
   // the caller that the user is currently scratching, and this is used to
@@ -52,24 +39,12 @@ public:
           bool* pReportScratching,
           bool* pReportReverse);
 
-  // Set rate change when temp rate button is pressed
-  static void setTemporaryRateChangeCoarseAmount(double v);
-  static double getTemporaryRateChangeCoarseAmount();
-  // Set rate change when temp rate small button is pressed
-  static void setTemporaryRateChangeFineAmount(double v);
-  static double getTemporaryRateChangeFineAmount();
   // Set rate change when perm rate button is pressed
   static void setPermanentRateChangeCoarseAmount(double v);
   static double getPermanentRateChangeCoarseAmount();
   // Set rate change when perm rate small button is pressed
   static void setPermanentRateChangeFineAmount(double v);
   static double getPermanentRateChangeFineAmount();
-  // Set Rate Ramp Mode
-  static void setRateRampMode(RampMode mode);
-  static RampMode getRateRampMode();
-  // Set Rate Ramp Sensitivity
-  static void setRateRampSensitivity(int);
-  static int getRateRampSensitivity();
   bool isReverseButtonPressed();
   // ReadAheadManager::getNextSamples() notifies us each time the play position
   // wrapped around during one buffer process (beatloop or track repeat) so
@@ -90,32 +65,13 @@ public slots:
   void slotControlFastBack(double);
 
 private:
-  void processTempRate(const size_t bufferSamples);
   double getJogFactor() const;
   double getWheelFactor() const;
   SyncMode getSyncMode() const;
 
-  // Set rate change of the temporary pitch rate
-  void setRateTemp(double v);
-  // Add a value to the temporary pitch rate
-  void addRateTemp(double v);
-  // Subtract a value from the temporary pitch rate
-  void subRateTemp(double v);
-  // Reset the temporary pitch rate
-  void resetRateTemp(void);
-  // Get the 'Raw' Temp Rate
-  double getTempRate(void);
-
   // Values used when temp and perm rate buttons are pressed
-  static ControlValueAtomic<double> m_dTemporaryRateChangeCoarse;
-  static ControlValueAtomic<double> m_dTemporaryRateChangeFine;
   static ControlValueAtomic<double> m_dPermanentRateChangeCoarse;
   static ControlValueAtomic<double> m_dPermanentRateChangeFine;
-
-  ControlPushButton* m_pButtonRateTempDown;
-  ControlPushButton* m_pButtonRateTempDownSmall;
-  ControlPushButton* m_pButtonRateTempUp;
-  ControlPushButton* m_pButtonRateTempUpSmall;
 
   ControlPushButton* m_pButtonRatePermDown;
   ControlPushButton* m_pButtonRatePermDownSmall;
@@ -146,6 +102,9 @@ private:
 
   ControlObject* m_pSampleRate;
 
+  // For pitch bending
+  PitchBendControl* m_pPitchBendControl;
+
   // For Sync Lock
   BpmControl* m_pBpmControl;
 
@@ -156,18 +115,8 @@ private:
   mixxx::audio::FramePos m_jumpPos;
   mixxx::audio::FramePos m_targetPos;
 
-  // This is true if we've already started to ramp the rate
-  bool m_bTempStarted;
-  // Set the Temporary Rate Change Mode
-  static RampMode m_eRateRampMode;
-  // The Rate Temp Sensitivity, the higher it is the slower it gets
-  static int m_iRateRampSensitivity;
   // Factor applied to the deprecated "wheel" rate value.
   static const double kWheelMultiplier;
   // Factor applied to jogwheels when the track is paused to speed up seeking.
   static const double kPausedJogMultiplier;
-  // Temporary pitchrate, added to the permanent rate for calculateRate
-  double m_tempRateRatio;
-  // Speed for temporary rate change
-  double m_dRateTempRampChange;
 };
